Only sync products that have been updated
We'll store a snapshot of data for every product in Airtable and use it
to figure out if the product needs to be synced: if the data has been
updated on Airtable it'll be different from the stored snapshot. In that
case we'll sync the product and update the snapshot.

This is obviously much more efficient than updating all products every
time.