Update images in WooCommerce only when necessary
Every time we pass images to WooCommerce as URLs it adds them to the
media library so its size keeps growing. We should therefore only pass
images when we know that they have changed.

It doesn't solve the issue completely, e.g. if a product has two images
and only one has been updated we'll pass 2 image URLs to WooCommerce and
it'll add both the updated and the old one to the media library again.
This is better than adding images to the library on every update though
and it would be quite tricky to keep track of updates to individual
images anyway.